Add files via upload
Could be used to get an idea of how to create camera angles. Update depending on the events that happen. Export as .pngs. In the scene there are three lines at the top right which is labeled as preview. You can create camera angles and access them as well as screenshot what's in the viewport. Holding ctrl and shift allows for moving objects on lower increments. Holding ctrl and selecting objects using the folders allow for you to modify multiple objects at the same time, modifying Locations and/or scale. Speaking of scale selecting an object/folder and clicking on transform at the top allows for you to scale objects up to a certain size.